requirements.txt
a2a-sdk==0.3.22
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
httpx>=0.27.0
requests>=2.31.0
